Override the engine libcrypto check in evp_extra_test
Unfortunately the evp_extra_test contains some tests
that access internal symbols and must therefore be
linked statically with libcrypto and recently some
engine tests were added, which triggers the newly
added engine load check.

It is pure luck that this did not crash so far.
If the same is done in 1.1.1 a crash happens immediately.
Likewise if the engine happens to call EVP_get_cipherbyname
or EVP_get_digestbyname in 3.0 or master.

As a work-around set an environment variable to override
the engine load check for now.