modify armv7_tick prototype to return uint64_t instead of uint32_t
Addresses an ABI issue on 32-bit arm.  The implementation of armv7_tick
calls the mrrc instruction to retrieve a counter from the p15 coprocessor.
This is a 64-bit counter and it is stored into the 2 32-bit registers
r0 and r1.  The problem is that the C prototype only declares the function
as returning a 32-bit value, so the compiler is not aware that the r1
register is being clobbered.  This results in undefined behavior and I
confirmed that I would get unpredictable results with a small test app
that duplicates this problem.  Modifying the return type to uint64_t
fixed the unpredictable results that I observed. I confirmed in the 32-bit
arm ABI documentation that a 64-bit result from a function is stored in
r0 and r1, which would inform the compiler that r1 is being overwritten
by the armv7_tick function properly.

This could be related to:

https://github.com/openssl/openssl/issues/14838
https://github.com/openssl/openssl/issues/17009
https://github.com/openssl/openssl/issues/17465

CLA: trivial