Adding documentation for X509_PURPOSE_add
documentation based on the problem descibed in detail in

https://github.com/openssl/openssl/issues/17959

Basically openssl ensures that if any eku is used eku_client has to be
used to establish proper connection. If that is not sufficient,
X509_PURPOSE_ANY has to be used.

the function X509_PURPOSE_add allows to define own purpose checking
function. Unfortunatelly (checked in depth in 1.1.1n) it is not possible
to use the function straight away, as x509_st is not well defined
(unless headers are copied). Also, some very useful functions are static
in v3_purp.c and not visible beyond.

This commit is mainly to start pull request discussion on how to deal
with the issue (if possible).