Fix various test failures with no-tls1_2 no-ec no-dh
Since no-tls1_2 implicitly disables tls1_1 and tls1
we have only tls1_3 available which is not usable
when no-ec and no-dh is used together.
Therefore we have zero usable tls protocols to test.
Fix that by removing "tls1_3" from avaliable_protocols,
because tls1_3 is not really available, even if it is
not explicitly disabled.