Fix cipherlist_test.c with no-tls1_2 no-tls1_3
When only TLS<1.2 is available and the security level is 0,
SSL_get1_supported_ciphers(ssl) retuns NULL.
That changed with commit 7bf2e4d7f0c7ae19b7a8c416910886a7171e9820.
Not sure of that is a bug or a feature, especially since
that does not happen with no-tls1_2 no-ec no-dh.
Test the current behaviour in case it changes again,
so we will know.