Resolve the BIO_ctrl() / OSSL_FUNC_BIO_ctrl_fn declaration clash
This is resolved by pretending that it isn't a clash, by making an internal
function in core that simply casts the result of `ossl_core_bio_ctrl()` to
`int` and hope for the best.  Since `OSSL_FUNC_BIO_ctrl_fn` declared an
`int` return value, that cast was done implicitly anyway.  Now it's done in
a more controlled manner.