Resolve the BIO_ctrl() / OSSL_FUNC_BIO_ctrl_fn declaration clash
This is resolved by pretending that it isn't a clash, by making an internal
function in core that simply casts the result of `ossl_core_bio_ctrl()` to
`int`.  Since `OSSL_FUNC_BIO_ctrl_fn` declared an `int` return value, that
cast was done implicitly anyway.  Now it's done in a more controlled manner.

The cast is guarded by an `ossl_assert()` that the result fits in an `int`,
so we notice in debug builds if a BIO ctrl ever returns a value that would
lose significant bits; in other builds, an error is returned instead of a
truncated value.

Assisted-by: Pi:moonshotai/kimi-k3
Signed-off-by: Richard Levitte <levitte@openssl.foundation>