Cygwin: Restore support for old versions (without `dladdr` and `Dl_info`).
But if a new enough version of Cygwin is detected,
we continue to take advantage of the presence of `dladdr` and `Dl_info`
(as per #9402 i.e. commit 38f6f99c).

"Detection" is based on diff's introduced in Cygwin on 2017-03-08
(`CYGWIN_VERSION_API_MINOR` bumped to `308` upon introduction of `dladdr`):
<https://cygwin.com/git/?p=newlib-cygwin.git;a=commit;h=c8432a01c8401c121940c806a9d868c4adc4cefd>.

Fixes #10200.

CLA: trivial